+++ packages/antd-management-fast-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.1.139
+
+### Patch Changes
+
+- update
+
 ## 1.1.138
 
 ### Patch Changes