+++ packages/antd-management-fast-develop/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.1.150
+
+### Patch Changes
+
+- update
+
 ## 1.1.149
 
 ### Patch Changes